[JSC][GreedyRegAlloc] Enable the greedy register allocator by default https://bugs.webkit.org/show_bug.cgi?id=288667 rdar://144792366
Reviewed by Mark Lam and Yusuke Suzuki.

Enable the greedy register allocator, except for wasm functions
that use SIMD, since SIMD is not yet supported.

The testPinRegisters case that's being disabled verifies that
the register allocator can coalesce a tmp with a pinned register.
That's not yet supported (it can coalesce with non-pinned fixed
registers, but not pinned registers). This optimization will be
implemented later. Note that this test is already disabled on ARM64.

* Source/JavaScriptCore/b3/testb3_7.cpp:
(testPinRegisters):
* Source/JavaScriptCore/runtime/OptionsList.h:

Canonical link: https://commits.webkit.org/291367@main