[UnifiedPDF] [iOS] Plugin missing when loading QLPreviewConverter-generated substituted data for supported MIME types https://bugs.webkit.org/show_bug.cgi?id=288869 rdar://145863043
Reviewed by Tim Horton.

For certain MIME types, we generate PDF representations on-the-fly using
QLPreviewConverter, and then use that as the substituted data when
writing the plugin document. Unfortunately, in this process, we end up
applying sandbox changes that restrict plugins, which disallows the
UnifiedPDFPlugin from being instantiated. This is driven through
Document::applyQuickLookSandbox().

To fix this, we loosen the plugin restriction on the QL sandbox in the
case the QL sandbox is being applied to a PluginDocument. This can only
happen when we want to present QL converted data in the PDF plugin.

Tested manually. API test will follow in a future patch.

* Source/WebCore/dom/Document.cpp:
(WebCore::Document::applyQuickLookSandbox):

Canonical link: https://commits.webkit.org/291372@main