[scroll-animations] WPT test scroll-animations/scroll-timelines/effect-updateTiming.html has failures https://bugs.webkit.org/show_bug.cgi?id=284542 rdar://141356805
Reviewed by Dean Jackson.

The Web Animations Level 1 specification indicates the following for the `duration` property
of the `OptionalEffectTiming` interface [0]:

    > The iteration duration, which is a real number greater than or equal to zero
    > (including positive infinity) representing the time taken to complete a single
    > iteration of the animation effect.

Looking through the Web Animations Level 2 and Scroll-driven Animations Level 1 specifications,
there is no text anywhere that indicates that an infinite duration for a progress-based animation
would yield an exception. As such, the test expecting `Infinity` for `duration` to throw is in
error. So we're modifying the relevant test to expect this to be a valid value instead.

Assuming there may have been an intent to restrict the `duration` for progress-based animations,
the following issue [1] was filed on the Scroll-driven Animations specification.

[0] https://drafts.csswg.org/web-animations-1/#dom-effecttiming-duration
[1] https://github.com/w3c/csswg-drafts/issues/11804

* LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/effect-updateTiming-expected.txt:
* LayoutTests/imported/w3c/web-platform-tests/scroll-animations/scroll-timelines/effect-updateTiming.html:

Canonical link: https://commits.webkit.org/291435@main