[Site Isolation] Remote page is sometimes not injected in to FrameProcess https://bugs.webkit.org/show_bug.cgi?id=288777 rdar://145796987
Reviewed by Charlie Wolfe.

In BrowsingContextGroup::addFrameProcessAndInjectPageContextIf, if the site associated with the
frame process matches an existing site for any pages in the browsing context group, then we bail out
of the function. This can leave us in a state where the FrameProcess does not have a RemotePage
associated with every page in the BCG.

No test, as it requires a transient state in which there's a page in the BCG for a given site and we
are also trying to create a new FrameProcess for that site at the same time (e.g. around unexpected
process termination time). In the usual case, we would just reuse the FrameProcess associated with
that site.

* Source/WebKit/UIProcess/BrowsingContextGroup.cpp:
(WebKit::BrowsingContextGroup::addFrameProcessAndInjectPageContextIf):

Canonical link: https://commits.webkit.org/291453@main