Reset the resumeTime and accumulated time on svg.setCurrentTime
https://bugs.webkit.org/show_bug.cgi?id=288906
rdar://145907044

Reviewed by Antoine Quint.

In 141765@main, we merged below Blink commit [1] but it was not complete fix.

[1] https://source.chromium.org/chromium/chromium/src/+/8c2c842380f44d69584abfb74633ef1049c37473

In [1] merge, we didn't added `lastResumeTime()` and it also led to issue where we were not
resetting the m_resumeTime and m_accumulatedActiveTime because of while the animation is paused
that time was getting accumulated if we unpaused the animation later. This bug is represented
by the test included in this merge [2].

[2] https://source.chromium.org/chromium/chromium/src/+/22866929d6e642fd4b64a37434507aab037c41de

* Source/WebCore/svg/animation/SMILTimeContainer.cpp:
(WebCore::SMILTimeContainer::elapsed const):
(WebCore::SMILTimeContainer::pause):
(WebCore::SMILTimeContainer::setElapsed):
* Source/WebCore/svg/animation/SMILTimeContainer.h:
* LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime.html:
* LayoutTests/svg/animations/getCurrentTime-pause-unpause-setCurrentTime-expected.txt:

Canonical link: https://commits.webkit.org/291486@main