Add Page Visibility API iframe Site Isolation Test https://bugs.webkit.org/show_bug.cgi?id=287569 rdar://problem/144710517
Reviewed by Alex Christensen.

Add test case to ensure that the page visibility api test works when in a remote iframe under site isolation.

On iOS, the `testRunner.setPageVisibility("hidden");` api is not working w/ or w/o site isolation, which
is causing the test to fail.

* LayoutTests/http/tests/site-isolation/page-visibility-onvisibilitychange-in-iframe-expected.txt: Added.
* LayoutTests/http/tests/site-isolation/page-visibility-onvisibilitychange-in-iframe.html: Added.
* LayoutTests/http/tests/site-isolation/resources/page-visibility-iframe-send-onvisibility-change-back.html: Added.
* LayoutTests/platform/ios/TestExpectations:

Canonical link: https://commits.webkit.org/291497@main