REGRESSION(290300@main): SaveAsPDF crashes if site-isolation is enabled https://bugs.webkit.org/show_bug.cgi?id=288620 rdar://145673736
Reviewed by Simon Fraser.

290300@main made the snapshot for site-isolation happens through a DisplayList.
But generating the PDFDocument was happening on the main thread. This caused a
threading issue which was fixed in 291031@main.

The fix of this threading issue in 291031@main was to move generating the PDFDocument
back to the `RemoteRendering WorkQueue` which is the correct approach.

But DrawingContext::replayDisplayList() requires a ControlFactory to be created
on the same thread. Otherwise it will call ControlFactory::shared() which creates
a MainThreadNeverDestroyed ControlFactory. This causes this crash because
ControlFactory::shared() has to be called only from the main thread.

The fix is to make RemoteRenderingBackend create a ControlFactory and pass it to
ImageBufferDisplayListBackend::create().

* LayoutTests/platform/mac-wk2/TestExpectations:
* Source/WebCore/platform/graphics/ImageBuffer.h:
(WebCore::ImageBuffer::create):
* Source/WebCore/platform/graphics/ImageBufferDisplayListBackend.cpp:
(WebCore::ImageBufferDisplayListBackend::create):
(WebCore::ImageBufferDisplayListBackend::ImageBufferDisplayListBackend):
(WebCore::ImageBufferDisplayListBackend::copyNativeImage):
(WebCore::ImageBufferDisplayListBackend::sinkIntoPDFDocument):
* Source/WebCore/platform/graphics/ImageBufferDisplayListBackend.h:
* Source/WebCore/platform/graphics/displaylists/DisplayListDrawingContext.cpp:
(WebCore::DisplayList::DrawingContext::replayDisplayList):
* Source/WebCore/platform/graphics/displaylists/DisplayListDrawingContext.h:
* Source/WebKit/GPUProcess/graphics/RemoteRenderingBackend.cpp:
(WebKit::allocateImageBufferInternal):
* Tools/TestWebKitAPI/Tests/WebKitCocoa/DrawingToPDF.mm:
(TestWebKitAPI::enableSiteIsolation):
(TestWebKitAPI::TEST(DrawingToPDF, SiteIsolationFormControl)):

Canonical link: https://commits.webkit.org/291521@main