[webkitbugspy] git-webkit fails with AttributeError if a tracked bug is not found when cc-ing https://bugs.webkit.org/show_bug.cgi?id=287517 rdar://144640536
Reviewed by Jonathan Bedard.

A possible edge case is when the Radar Importer is CC'd on a bug but it is slow to respond
and there is no radar associated with the bug. Then, a user tries to manually CC a radar through
git-webkit. Our tooling will believe there's already a radar associated when there isn't.

We shouldn't output '?' in place of a radar and rather catch this case and verify
that the user intended to CC the radar. (Another possibility is to silently CC the radar without
asking, since this is technically correct if we discount race conditions with the importer.)

* Tools/Scripts/libraries/webkitbugspy/webkitbugspy/bugzilla.py:
(Tracker.cc_radar):

Canonical link: https://commits.webkit.org/291529@main