[JSC] Continue test coverage of the graph coloring register allocator https://bugs.webkit.org/show_bug.cgi?id=288973 rdar://146025608
Reviewed by Yusuke Suzuki.

Until the graph coloring register allocator is removed (rdar://144792585),
let's continue to get some test coverage of it.

In order to not increase stress test time too much, for JS tests, I've
sacrificed a couple of configurations that seem not super critical to run
with all allocators and run those with the graph allocator. For wasm,
given how the configs are organized, there didn't seem like a great
way to do that so I have added another test config for the graph allocator,
but wasm is a much smaller set of tests so this seems okay.

I'm holding off on removing --airForce{Briggs,IRC}Allocator options
from configurations that no longer use the graph allocator so that this
change is not dependent on 291367@main, in case that needs to be reverted
for whatever reason. These options will be removed from JSC and scrubbed
here when the graph coloring allocators are removed.

* Tools/Scripts/run-jsc-stress-tests:

Canonical link: https://commits.webkit.org/291547@main