[WebGPU] Shader validation routines potentially rely on UB in C++ https://bugs.webkit.org/show_bug.cgi?id=288726 rdar://145760120
Reviewed by Cameron McCormack.

The shader was using a buffer of sizeof(S) + sizeof(uint)
and writing to &S::lastUintMember + 1 assuming it would
write to the uint.

This is flagged by sanitizers, so be explict and define
the member we are writing to.

* Source/WebGPU/WebGPU/Buffer.h:
* Source/WebGPU/WebGPU/Buffer.mm:
(WebGPU::Buffer::Buffer):
(WebGPU::Buffer::makeIndexIndirectBuffer): Deleted.
* Source/WebGPU/WebGPU/Device.h:
* Source/WebGPU/WebGPU/Device.mm:
* Source/WebGPU/WebGPU/RenderPassEncoder.mm:
(WebGPU::RenderPassEncoder::clampIndexBufferToValidValues):
(WebGPU::checkForIndirectDrawDeviceLost):

Canonical link: https://commits.webkit.org/291576@main