Use smart pointers in WebXROpaqueFramebuffer https://bugs.webkit.org/show_bug.cgi?id=285629 rdar://122475331
Reviewed by Dan Glastonbury.

WebXROpaqueFramebuffer stored a C++ reference to the WebGL context,
but it is not clear who ensures the GL context is retained during
the lifetime of the WebXROpaqueFramebuffer.

Via code inspection, the out of bounds access should only occur
if the reference was invalid or, more likely, WebGLRenderingContextBase::initializeContextState()
was never called, in which case the store / resolve should be a nop.

* Source/WebCore/Modules/webxr/WebXROpaqueFramebuffer.cpp:
(WebCore::WebXROpaqueFramebuffer::WebXROpaqueFramebuffer):
(WebCore::WebXROpaqueFramebuffer::~WebXROpaqueFramebuffer):
(WebCore::WebXROpaqueFramebuffer::startFrame):
(WebCore::WebXROpaqueFramebuffer::endFrame):
Bail if the WeakPtr is null.

* Source/WebCore/Modules/webxr/WebXROpaqueFramebuffer.h:
Store a WeakPtr instead of a &.

* Source/WebCore/html/canvas/WebGLUtilities.h:
(WebCore::ScopedWebGLRestoreTexture::~ScopedWebGLRestoreTexture):
Ensure we don't access out of bounds in the texture units.

Canonical link: https://commits.webkit.org/291611@main