CRASH in VideoPresentationInterfaceIOS::ensurePipPlacardIsShowing rdar://146108129 https://bugs.webkit.org/show_bug.cgi?id=289107
Reviewed by Brent Fulgham.

Crash logging shows an exception thrown from +[NSArray arrayWithObjects:count:] within
ensurePipPlacardIsShowing(), which can only be caused by a nil object being inserted into
an objective-c literal NSArray. There are a couple of code paths which generate a literal
array in that method, but given that the failure case is that the plackard
informing the user that the video is in PiP will not be shown, we should probably not
crash here. Wrap the method in a try/catch block.

* Source/WebCore/platform/ios/VideoPresentationInterfaceIOS.mm:
(WebCore::VideoPresentationInterfaceIOS::ensurePipPlacardIsShowing):

Canonical link: https://commits.webkit.org/291622@main