Add support for generating a consumer for CSS property 'background-clip' https://bugs.webkit.org/show_bug.cgi?id=288855
Reviewed by Antti Koivisto.

Supporting 'background-clip' required adding support for annotating terms
in the grammar with a settings flag. The syntax uses an @-annotation:

     `[ foo bar@(settings-flag=cssBarEnabled) baz]`

To make this work, support for setting annotations on keywords was needed.
This proved a bit challenging with the parser, so a bit of refactoring to
explicitly track the nodes multipliers and annotations will be attached to
was added.

There were also a number of spelling issues in process-css-properties.py
that were corrected.

* Source/WebCore/css/CSSProperties.json:
* Source/WebCore/css/parser/CSSPropertyParser.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Background.cpp:
* Source/WebCore/css/parser/CSSPropertyParserConsumer+Background.h:
* Source/WebCore/css/process-css-properties.py:

Canonical link: https://commits.webkit.org/291625@main