Fix invalid WeakRef deference crash in WebCore's Memory Pressure Handler https://bugs.webkit.org/show_bug.cgi?id=289117 rdar://144487022
Reviewed by Ryan Reno, Cameron McCormack, and Matthieu Dubet.

We received crash reports of an invalid access to the document
WeakRef while iterating through the allDocuments() map in releaseNoncriticalMemory.
It is not clear what is causing the document to be deleted, perhaps
related to clearing async node deletion list . Although, this should not be possible due
to excluding iframes to be added to the deletion queue and dissallowing Documents
and DocumentFragments to have their immediate children be async deleted.

A fix to this specific crash is to create a list of Ref<Document> from the allDocuments() map,
and iterate through this list.

Change the WeakRef<Document> used to call member functions on the
document into a Ref<Document>, and ensure that the DocumentsMap is not being
modified while clearing the various data structures that the Document holds.

Now, no documents in the map can be deleted while in this function.

* Source/WebCore/page/MemoryRelease.cpp:
(WebCore::releaseNoncriticalMemory):

Canonical link: https://commits.webkit.org/291644@main