[iOS] [Visual Bidi Selection] Selection should be visually contiguous across soft line breaks in bidi text https://bugs.webkit.org/show_bug.cgi?id=289184 rdar://142811818
Reviewed by Abrar Rahman Protyasha.

Make some adjustments to the visual bidi selection mode on iOS. See below for more details.

* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-8-expected.txt: Added.
* LayoutTests/editing/selection/ios/bidi-visually-contiguous-selection-8.html: Added.

Add a layout test to exercise this change, by extending a selection handle across a soft line break
containing bidi LTR text in an RTL paragraph, and verify that the selection snaps to select the
entire LTR text run after text interaction ends.

* LayoutTests/resources/ui-helper.js:
(window.UIHelper.async selectionHitTestOffset):
(window.UIHelper.async zoomByDoubleTappingAt): Deleted.
* Source/WebCore/editing/Editing.cpp:
(WebCore::advanceInDirection):
(WebCore::makeVisuallyContiguousIfNeeded):

Consider the case where, in an RTL paragraph with selected text, the selection end position is in
the second line of an LTR text run that spans two lines with a soft line break in between. Right
now, we treat this scenario as being visually contiguous and don't attempt to adjust the selection
endpoints because there's no break in visual continuity in the selection, within the last line.
However, this can lead to unintuitive behaviors, as the selection will visually flip when extended
across the bidi text boundary, to the rest of the RTL text on the second line. To mitigate this, we
detect the following case:

-   The selection spans multiple lines.
-   The visually selected text on the first or last line only spans text which is a different
    direction from the rest of the paragraph (in this case, only LTR text is selected when the rest
    of the paragraph is RTL).

...and automatically expand the selection start or end, so that we select all of the text that's of
a different direction than the paragraph direction after the soft line break.

* Source/WebCore/rendering/RenderObject.cpp:
(WebCore::makeBidiSelectionVisuallyContiguousIfNeeded):
(WebCore::adjustTextDirectionForCoalescedGeometries):

Adjust code that makes iOS selection rects visually contiguous, so that it will start the selection
rect from the right visual extent in the scenario detailed above. To achieve this, we compute the
visible extent of the first or last line of the selection, on the opposite end of the caret, and
unify the visual selection bounds with this caret rect in order to compute the final coalesced quad
for the start and end selection geometries.

(WebCore::directionForFirstLine): Deleted.
(WebCore::directionForLastLine): Deleted.

Canonical link: https://commits.webkit.org/291662@main